# server/openapi.json
{
  "components": {
    "schemas": {
...
      "ChaseActionRequest": {
        "description": "一轮追逐行动。type: run（奔逃）| hazard（闯障，附 hazard 明细）。",
        "properties": {
          "hazard": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hazard"
          },
          "type": {
            "default": "run",
            "title": "Type",
            "type": "string"
          }
        },
        "title": "ChaseActionRequest",
        "type": "object"
      },
...
      "CombatActionRequest": {
        "description": "一个战斗行动。type: attack | dodge | fight_back | flee | first_aid | observe |\nmaneuver | reload | aim | other。",
        "properties": {
          "defense": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defense"
          },
          "dest": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dest"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "shots": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shots"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "type": {
            "default": "attack",
            "title": "Type",
            "type": "string"
          },
          "weapon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weapon"
          }
        },
        "title": "CombatActionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ParticipantInput": {
        "description": "开局时前端提交的一个席位。\n\ncharacter_id 为空且 role=human 表示「留空待加入」的空席（claimed=false）。",
        "properties": {
          "character_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Id"
          },
          "is_primary": {
            "default": false,
            "title": "Is Primary",
            "type": "boolean"
          },
          "role": {
            "default": "ai",
            "title": "Role",
            "type": "string"
          }
        },
        "title": "ParticipantInput",
        "type": "object"
      },
...
      "ReactionRequest": {
        "description": "被攻击时的反应。choice: fight_back | dodge | cover。",
        "properties": {
          "choice": {
            "title": "Choice",
            "type": "string"
          }
        },
        "required": [
          "choice"
        ],
        "title": "ReactionRequest",
        "type": "object"
      },
...
      "SessionCreate": {
        "properties": {
          "kp_mode": {
            "default": "ai",
            "title": "Kp Mode",
            "type": "string"
          },
          "module_id": {
            "title": "Module Id",
            "type": "string"
          },
          "participants": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ParticipantInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participants"
          },
          "player_character_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Character Id"
          }
        },
        "required": [
          "module_id"
        ],
        "title": "SessionCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Player-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "TRPG Player",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/sessions": {
      "get": {
        "operationId": "list_sessions_api_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "sessions"
        ]
      },
      "post": {
        "operationId": "create_session_api_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Session",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/api/sessions/{session_id}/chase": {
      "get": {
        "description": "当前追逐态（无则 {active:false}），供前端渲染距离轨与重连对齐。",
        "operationId": "get_chase_api_sessions__session_id__chase_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Chase",
        "tags": [
          "combat"
        ]
      }
    },
    "/api/sessions/{session_id}/chase/action": {
      "post": {
        "description": "玩家推进一轮追逐（奔逃/闯障）。结算后广播 dice/chase/system chunks。",
        "operationId": "chase_action_api_sessions__session_id__chase_action_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChaseActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Chase Action",
        "tags": [
          "combat"
        ]
      }
    },
...
    "/api/sessions/{session_id}/combat": {
      "get": {
        "description": "当前战斗态（无则 {active:false}），供前端渲染战斗视图与断线重连对齐。",
        "operationId": "get_combat_api_sessions__session_id__combat_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Combat",
        "tags": [
          "combat"
        ]
      }
    },
    "/api/sessions/{session_id}/combat/action": {
      "post": {
        "description": "提交一个战斗行动（attack/dodge/fight_back/flee/first_aid/observe/maneuver/reload/aim/other）。\n仅当轮到本玩家时有效。",
        "operationId": "combat_action_api_sessions__session_id__combat_action_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CombatActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Combat Action",
        "tags": [
          "combat"
        ]
      }
    },
    "/api/sessions/{session_id}/combat/reaction": {
      "post": {
        "description": "被攻击的真人提交反应（fight_back/dodge/cover）。token 对应的角色即防御者。\n\n结算这一击并续跑驱动，广播新的 combat_state/dice/narration。",
        "operationId": "combat_reaction_api_sessions__session_id__combat_reaction_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Combat Reaction",
        "tags": [
          "combat"
        ]
      }
    },
    "/api/sessions/{session_id}/combat/roll": {
      "post": {
        "description": "两段式攻击第二段：玩家亲自掷伤害。仅当有属于本玩家的 pending_roll 时有效。",
        "operationId": "combat_roll_api_sessions__session_id__combat_roll_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Combat Roll",
        "tags": [
          "combat"
        ]
      }
    },
...
  }
}